+++ tests/testthat/test-api.R
@@ -43,7 +43,7 @@ test_that("retrieving persons for KTH projects works", {
     purrr::pluck(1)
 
   kthp <- swecris_persons(kthid)
-  is_valid <- nrow(kthp) > 500
+  is_valid <- nrow(kthp) > 300
   expect_true(is_valid)
 
 })
@@ -79,7 +79,7 @@ test_that("Fundings for KTH project works", {
 
 test_that("Projects data can be retrieved from an ORCiD", {
   o <- "0000-0003-1102-4342" %>% swecris_projects_from_orcid()
-  is_valid <- nrow(o$projects) > 10 & nrow(o$peopleList) > 100 & nrow(o$scbs) > 10
+  is_valid <- nrow(o$projects) > 10 & nrow(o$peopleList) > 40 & nrow(o$scbs) > 10
   expect_true(is_valid)
 })
 
@@ -91,7 +91,7 @@ test_that("A single project can be retrieved from a project identifier", {
 
 test_that("A single project's associated people can be retrieved from a project identifier", {
   res <- "2021-00157_VR" %>% swecris_project_people()
-  is_valid <- ncol(res) > 4 & nrow(res) > 2
+  is_valid <- ncol(res) > 4 & nrow(res) == 1
   expect_true(is_valid)
 })
 
@@ -100,3 +100,5 @@ test_that("A single project's associated SCB codes can be retrieved from a proje
   is_valid <- ncol(res) > 2 & nrow(res) > 1
   expect_true(is_valid)
 })
+
+